+++ resources/views/email/raw.blade.php
@@ -1,0 +1,2 @@
+{{-- Verbatim passthrough for resending stored, already-rendered emails. --}}
+{!! $body !!}